refactor(rr-lib): separate templates
