+++ apps/react-router-lib/src/routes/about/route.tsx
@@ -1,8 +1,6 @@
 import { DocumentTitle } from '@learn-react/core/src/components/utils/DocumentTitle';
 import { type FC } from 'react';
-import { Link } from 'react-router';
-import { Paths } from '../../routes';
-import styles from './styles.module.css';
+import { Template } from './template';
 
 /**
  * The about page.
@@ -11,34 +9,6 @@ export const Component: FC = () => (
   <>
     <DocumentTitle title="About React Router Contacts" />
 
-    <div className={styles.base}>
-      <Link viewTransition to={Paths.Home}>
-        ← Go to demo
-      </Link>
-
-      <h1>About React Router Contacts</h1>
-
-      <div>
-        <p>
-          This is a demo application showing off some of the powerful features of React Router, including dynamic
-          routing, nested routes, loaders, actions, and more.
-        </p>
-
-        <h2>Features</h2>
-        <p>Explore the demo to see how React Router handles:</p>
-        <ul>
-          <li>Data loading and mutations with loaders and actions</li>
-          <li>Nested routing with parent/child relationships</li>
-          <li>URL-based routing with dynamic segments</li>
-          <li>Pending and optimistic UI</li>
-        </ul>
-
-        <h2>Learn More</h2>
-        <p>
-          Check out the official documentation at <a href="https://reactrouter.com">reactrouter.com</a> to learn more
-          about building great web applications with React Router.
-        </p>
-      </div>
-    </div>
+    <Template />
   </>
 );
