+++ apps/react-router-lib/src/routes/contacts/contact/route.tsx
@@ -1,11 +1,8 @@
 import { DocumentTitle } from '@learn-react/core/src/components/utils/DocumentTitle';
-import { type FC, type FormEvent } from 'react';
-import { Form, generatePath, useLoaderData, type Params } from 'react-router';
-import { Button } from '../../../components/Button';
+import { type FC } from 'react';
+import { useLoaderData, type Params } from 'react-router';
 import { getContact, updateContact } from '../../../data';
-import { Paths } from '../../../routes';
-import { Favorite } from './Favorite';
-import styles from './styles.module.css';
+import { Template } from './template';
 
 /**
  * Fetches contact data and returns it as a prop.
@@ -53,74 +50,11 @@ export async function action({ request, params: { contactId } }: { request: Requ
 export const Component: FC = () => {
   const { contact } = useLoaderData<typeof loader>();
 
-  const handleDestroy = (event: FormEvent) => {
-    // eslint-disable-next-line no-alert, no-restricted-globals
-    const response = confirm('Please confirm you want to delete this record.');
-
-    if (!response) {
-      event.preventDefault();
-    }
-  };
-
   return (
     <>
       <DocumentTitle title={`${contact.first ?? ''} ${contact.last ?? ''}`} />
 
-      <div className={styles.base}>
-        <div>
-          <img
-            key={contact.avatar}
-            className={styles.image}
-            alt={`${contact.first ?? ''} ${contact.last ?? ''} avatar`}
-            src={contact.avatar}
-          />
-        </div>
-
-        <div>
-          <h1 className={styles.title}>
-            {contact.first || contact.last ? (
-              <>
-                {contact.first} {contact.last}
-              </>
-            ) : (
-              <i>No Name</i>
-            )}
-            <Favorite favorite={contact.favorite} />
-          </h1>
-
-          {contact.twitter ? (
-            <p className={styles.twitter}>
-              <a className={styles.link} href={`https://x.com/${contact.twitter}`}>
-                {contact.twitter}
-              </a>
-            </p>
-          ) : null}
-
-          {contact.notes ? <p className={styles.notes}>{contact.notes}</p> : null}
-
-          <div className={styles.controls}>
-            <Form
-              action={generatePath(Paths.Contacts.Edit, {
-                contactId: contact.id,
-              })}
-            >
-              <Button type="submit">Edit</Button>
-            </Form>
-
-            <Form
-              action={generatePath(Paths.Contacts.Destroy, {
-                contactId: contact.id,
-              })}
-              method="post"
-              onSubmit={handleDestroy}
-            >
-              <Button theme="danger" type="submit">
-                Delete
-              </Button>
-            </Form>
-          </div>
-        </div>
-      </div>
+      <Template contact={contact} />
     </>
   );
 };
