+++ apps/react-router-lib/src/routes/contacts/edit/route.tsx
@@ -1,13 +1,9 @@
 import { DocumentTitle } from '@learn-react/core/src/components/utils/DocumentTitle';
-import { clsx } from 'clsx';
 import { type FC } from 'react';
-import { Form, generatePath, redirect, useLoaderData, useNavigate, type Params } from 'react-router';
-import { Button } from '../../../components/Button';
-import { TextArea } from '../../../components/TextArea';
-import { TextInput } from '../../../components/TextInput';
+import { generatePath, redirect, useLoaderData, useNavigate, type Params } from 'react-router';
 import { getContact, updateContact } from '../../../data';
 import { Paths } from '../../../routes';
-import styles from './styles.module.css';
+import { Template } from './template';
 
 /**
  * Fetches contact data and returns it as a prop.
@@ -70,51 +66,7 @@ export const Component: FC = () => {
     <>
       <DocumentTitle title={`Edit ${contact.first} ${contact.last}`} />
 
-      <Form key={contact.id} className={styles.base} method="post">
-        <p className={styles['form-group']}>
-          <span className={styles['label-display']}>Name</span>
-          <div className={styles['form-item']}>
-            <TextInput aria-label="First name" defaultValue={contact.first} name="first" placeholder="First" />
-            <TextInput aria-label="Last name" defaultValue={contact.last} name="last" placeholder="Last" />
-          </div>
-        </p>
-
-        {}
-        <label className={styles['form-group']}>
-          <span className={styles['label-display']}>Twitter</span>
-          <div className={styles['form-item']}>
-            <TextInput name="twitter" defaultValue={contact.twitter} placeholder="@jack" />
-          </div>
-        </label>
-
-        {}
-        <label className={styles['form-group']}>
-          <span className={styles['label-display']}>Avatar URL</span>
-          <div className={styles['form-item']}>
-            <TextInput
-              aria-label="Avatar URL"
-              defaultValue={contact.avatar}
-              name="avatar"
-              placeholder="https://example.com/avatar.jpg"
-            />
-          </div>
-        </label>
-
-        {}
-        <label className={styles['form-group']}>
-          <span className={styles['label-display']}>Notes</span>
-          <div className={styles['form-item']}>
-            <TextArea defaultValue={contact.notes} name="notes" rows={6} />
-          </div>
-        </label>
-
-        <p className={clsx(styles['form-group'], styles.controls)}>
-          <Button>Save</Button>
-          <Button type="button" onClick={handleCancel}>
-            Cancel
-          </Button>
-        </p>
-      </Form>
+      <Template contact={contact} onCancel={handleCancel} />
     </>
   );
 };
