+++ apps/react-router-lib/src/routes/home/route.tsx
@@ -1,6 +1,6 @@
 import { DocumentTitle } from '@learn-react/core/src/components/utils/DocumentTitle';
 import { type FC } from 'react';
-import styles from './styles.module.css';
+import { Template } from './template';
 
 /**
  * The Home page.
@@ -9,14 +9,6 @@ export const Component: FC = () => (
   <>
     <DocumentTitle title="React Router Contacts" />
 
-    <p className={styles.base}>
-      This is a demo for React Router.
-      <br />
-      Check out{' '}
-      <a href="https://reactrouter.com" className={styles.link}>
-        the docs at reactrouter.com
-      </a>
-      .
-    </p>
+    <Template />
   </>
 );
