+++ package.json
@@ -42,7 +42,7 @@
     "lint-staged": "15.3.0",
     "prettier": "3.4.2",
     "prettier-plugin-organize-imports": "4.1.0",
-    "stylelint": "16.13.0",
+    "stylelint": "16.12.0",
     "typescript": "5.7.3"
   }
 }
